user-service/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
alembic==1.10.3
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.0.0
asyncpg==0.27.0
aiosmtplib==2.0.1
jinja2==3.1.2
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0